backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
google-generativeai==0.5.3
pvcheetah==2.1.1
python-dotenv==1.0.1
